requirements_test.txt
pytest
pytest-asyncio
pytest-cov
pytest-homeassistant-custom-component
mypy
ruff
python-homely==0.1.2
